- Check for admin and suspended status.
- Added debugbar.
- Removed policies folder, makes project bulky and there is no real need for it right now.
- Disable Telescope and Pulse by default.
- Fixed favicon throwing 404.

+++ app/Http/Controllers/ServerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\Server\StoreServerRequest;
 use App\Http\Requests\Server\UpdateServerRequest;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Server;
 
 class ServerController extends Controller
@@ -14,22 +15,6 @@ class ServerController extends Controller
     public function index()
     {
         return view('server.index', ['servers' => Server::all()]);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(StoreServerRequest $request)
-    {
-        //
     }
 
     /**
@@ -47,6 +32,7 @@ class ServerController extends Controller
      */
     public function edit(Server $server)
     {
+        Gate::authorize('admin-only');
         return view('server.edit', [
             'server' => Server::find($server->server_id)
         ]);
@@ -57,10 +43,10 @@ class ServerController extends Controller
      */
     public function update(UpdateServerRequest $request, Server $server)
     {
+        Gate::authorize('admin-only');
+
         $server->fill($request->validated());
-
         $server->save();
-
         return to_route('server.edit', $server->server_id)->with('status', 'server-updated');
     }
 
@@ -69,8 +55,9 @@ class ServerController extends Controller
      */
     public function destroy(Server $server)
     {
-        $server->delete();
+        Gate::authorize('admin-only');
 
+        $server->delete();
         return to_route('server.index');
     }
 }
